Remove unnecessary ProtocolUtils extraction
